requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.7.2
jp==1.8.0
langchain==0.2.14
langchain-chroma==0.1.3
langchain-community==0.2.12
langchain-core==0.2.34
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.2
python-dotenv==1.0.1
pypdf==4.3.1